% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{cmp}
\alias{cmp}
\title{Title}
\usage{
cmp(n.rep = 100, X, Y, train_prop = 0.9)
}
\arguments{
\item{n.rep}{numeric}

\item{X}{matrix}

\item{Y}{matrix}

\item{train_prop}{numeric}
}
\value{
list
}
\description{
Title
}
\examples{
\dontrun{
cmp(X = HAM$hamimgs, Y = HAM$hamlab)
}
}
