% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_srrmlr.R
\name{cv_srrmlr}
\alias{cv_srrmlr}
\title{Multinomial logistic regression model with SRRMLR}
\usage{
cv_srrmlr(X, Y, train_prop = 0.9)
}
\arguments{
\item{X}{matrix}

\item{Y}{matrix}

\item{train_prop}{numeric}
}
\value{
list
}
\description{
Construct a regression model with responses be multinomial variable,
predictors be continuous high-dimensional data. In addition, this function
provides results obtained from other 4 baseline methods.
}
\examples{
\dontrun{
cv_srrmlr(HAM$hamimgs, HAM$hamlab)
}
}
