% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{gendata}
\alias{gendata}
\title{Generate data for simulation}
\usage{
gendata(n, p, q = 8, s = 5, r = 5, type, C = NULL)
}
\arguments{
\item{n}{number of observations}

\item{p}{dimension of predictor vector}

\item{q}{dimension of response vector}

\item{s}{sparsity level of the coefficient matrix}

\item{r}{the specified rank}

\item{type}{'AR' for Auto Regressive, 'CS' for Compound Symmetries}

\item{C}{the true coefficient matrix}
}
\value{
Y generated response matrix

X generated covariate matrix

C generated true coefficient matrix

B generated B matrix

V generated V matrix
}
\description{
Generate data for simulation
}
