\name{SRRMLR-package}
\alias{SRRMLR-package}
\alias{SRRMLR}
\docType{package}
\title{\packageTitle{SRRMLR}}
\description{\packageDescription{SRRMLR}}
\section{Package Content}{\packageIndices{SRRMLR}}
\author{\packageAuthor{SRRMLR}}
\section{Maintainer}{\packageMaintainer{SRRMLR}}
\keyword{package}
