% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{srrmlr_one}
\alias{srrmlr_one}
\title{The base function of SRRMLR}
\usage{
srrmlr_one(X, Y, s, r, e_in = 0.001, e_out = 0.001)
}
\arguments{
\item{X}{the covariate matrix}

\item{Y}{the response matrix}

\item{s}{the sparsity}

\item{r}{the rank}

\item{e_in}{the stopping epsilon of inloop}

\item{e_out}{the stopping epsilon of outloop}
}
\value{
C matrix aims to be solved

iters iterations taken when algorithm meet the stopping criterion

Prob the probability matrix
}
\description{
The base function of SRRMLR
}
