% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{srrmlr_real}
\alias{srrmlr_real}
\title{SRRMLR in real data analysis}
\usage{
srrmlr_real(train_X, train_Y, test_X, test_Y, s_list, r_list, cv = 5L)
}
\arguments{
\item{train_X}{the covariate matrix of training data}

\item{train_Y}{the response matrix of training data}

\item{test_X}{the covariate matrix of test data}

\item{test_Y}{the response matrix of test data}

\item{s_list}{the list of sparsity}

\item{r_list}{the list of rank}

\item{cv}{the number of folds when conducting cross-validation, 5 in default}
}
\value{
Pred the predictive accuracy with the negative log-likelihood

r rank selected by SRRMLR, the estimated rank

s sparsity selected by SRRMLR, also the number of non-zero rows in C_hat

A_set the index of features selected by SRRMLR.

time running time of the algorithm with selected rank and sparsity
}
\description{
SRRMLR in real data analysis
}
