% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{srrmlr_simu}
\alias{srrmlr_simu}
\title{SRRMLR in simulation analysis}
\usage{
srrmlr_simu(
  train_X,
  train_Y,
  val_X,
  val_Y,
  test_X,
  test_Y,
  s_list,
  r_list,
  C_true,
  s_true,
  r_true
)
}
\arguments{
\item{train_X}{the covariate matrix of training data}

\item{train_Y}{the response matrix of training data}

\item{val_X}{the covariate matrix of validation data}

\item{val_Y}{the response matrix of validation data}

\item{test_X}{the covariate matrix of test data}

\item{test_Y}{the response matrix of test data}

\item{s_list}{the list of sparsity}

\item{r_list}{the list of rank}

\item{C_true}{the true matrix C}

\item{s_true}{the true sparsity}

\item{r_true}{the true rank}
}
\value{
Pred the predictive accuracy with the negative log-likelihood

Est the estimation accuracy

r rank selected by SRRMLR, the estimated rank

s sparsity selected by SRRMLR, also the number of non-zero rows in C_hat

Sen the sensitivity

Spe the specificity

time running time of the algorithm with selected rank and sparsity
}
\description{
SRRMLR in simulation analysis
}
